Added RecentlyNotified and Notified methods to Subscriptions
RecentlyNotified tells us if the subscription caused a recent
notification. This ensures we dont spam users with notifications.

Notified simply adds an event singalling the subscription
was used.

Tests added for this new functionality. We really need to
setup a proper testing db as all testing ids differ between
services and as such functional code will fail depening on the
db data.