Add SubscribedForToday method to subscription
Needs a test, however testing this is a bit clunky. It would
either require modifying the method signature to take either a custom
clock implementation allowing us to fake the date or nil to use
the real clock. Or injecting with a similar implementation. Or
finally using some other package that allows patching the time.Now func.

Either way, takes time we don't have atm.